EXP 1/requirements.txt
sentence-transformers
numpy
faiss-cpu
langchain_openai
langchain_core
langgraph
streamlit
torch
python-dotenv